# specs/openapi.yaml
openapi: 3.1.0
info:
  title: Counter POS V1 API
  version: 1.0.0
  description: Implemented local single-store API. Same-origin cookie and CSRF
    authentication. Cash-only offline; external methods are manual records.
    Runtime validation and financial invariants also require the shared domain
    rules. See docs/adr/001-v1-implementation.md.
servers:
  - url: /api/v1
    description: Same-origin application API
security:
  - cookieAuth: []
paths:
  /auth/login:
    post:
      summary: Sign in online
      operationId: post__auth_login
      description: Validate Origin and login rate limits. Set secure HttpOnly cookie;
        rotate session. Never return the cookie as JSON.
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/Unavailable"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Login"
      security: []
  /auth/logout:
    post:
      summary: Revoke server session
      operationId: post__auth_logout
      description: Revoke server session
      responses:
        "204":
          description: Successful operation
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/Unavailable"
      security:
        - cookieAuth: []
          csrfToken: []
...
  /demo/login:
    post:
      summary: Local synthetic demo session; unavailable in production
      description: Local synthetic demo session; unavailable in production
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  type: string
                  enum:
                    - MANAGER
                    - CASHIER
              required:
                - role
              additionalProperties: false
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
  /users:
    get:
      summary: List store operators
      description: Manager only; normal cookie/CSRF authorization.
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Store operators; credentials excluded
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      summary: Create store operator
      description: Manager only; normal cookie/CSRF authorization.
      security:
        - cookieAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUser"
      responses:
        "201":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    patch:
      summary: Change operator access with optimistic version; revoke old sessions
      description: Manager only; normal cookie/CSRF authorization.
      security:
        - cookieAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUser"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      parameters:
        - name: id
          in: path
          required: true
          schema: *a1
  /users/{id}/password:
    post:
      summary: Reset another operator's password and revoke sessions
      description: Manager only; normal cookie/CSRF authorization.
      security:
        - cookieAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPassword"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      parameters:
        - name: id
          in: path
          required: true
          schema: *a1
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: counter_session
      description: Secure HttpOnly SameSite session cookie set by server.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Authenticated anti-CSRF token; validate Origin too.
  schemas:
    Error:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 80
        message:
          type: string
          minLength: 1
          maxLength: 500
        request_id:
          type: string
          minLength: 1
          maxLength: 100
        retryable:
          type: boolean
        details:
          type: object
          additionalProperties: true
      required:
        - code
        - message
        - request_id
        - retryable
    User:
      type: object
      properties:
        id: *a1
        email:
          type: string
          format: email
        display_name:
          type: string
        role:
          type: string
          enum:
            - MANAGER
            - CASHIER
        active:
          type: boolean
        version:
          type: integer
          minimum: 1
        created_at:
          type: string
          format: date-time
      required:
        - id
        - email
        - display_name
        - role
        - active
        - version
        - created_at
      additionalProperties: false
    Login:
      type: object
      additionalProperties: false
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          format: password
          minLength: 1
          maxLength: 200
          writeOnly: true
      required:
        - email
        - password
    LoginResult:
      type: object
      properties:
        user:
          type: object
        csrf_token:
          type: string
      required:
        - user
        - csrf_token
      additionalProperties: false
...
    CreateUser:
      type: object
      properties:
        email:
          type: string
          maxLength: 254
          format: email
          pattern: ^(?:[A-Za-z0-9_'+\-]+\.)*[A-Za-z0-9_'+\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\-]*\.)+[A-Za-z]{2,}$
        display_name:
          type: string
          minLength: 1
          maxLength: 100
        role:
          type: string
          enum:
            - MANAGER
            - CASHIER
        password:
          type: string
          minLength: 12
          maxLength: 200
      required:
        - email
        - display_name
        - role
        - password
      additionalProperties: false
    UpdateUser:
      type: object
      properties:
        version:
          type: integer
          exclusiveMinimum: 0
          maximum: 9007199254740991
        display_name:
          type: string
          minLength: 1
          maxLength: 100
        role:
          type: string
          enum:
            - MANAGER
            - CASHIER
        active:
          type: boolean
        reason:
          type: string
          minLength: 1
          maxLength: 500
      required:
        - version
        - reason
      additionalProperties: false
    ResetPassword:
      type: object
      properties:
        version:
          type: integer
          exclusiveMinimum: 0
          maximum: 9007199254740991
        new_password:
          type: string
          minLength: 12
          maxLength: 200
        reason:
          type: string
          minLength: 1
          maxLength: 500
      required:
        - version
        - new_password
        - reason
      additionalProperties: false
...
  responses:
    BadRequest:
      description: Malformed/invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Role or scope denied
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Idempotency/version/stock conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unprocessable:
      description: Business rule or review required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooManyRequests:
      description: Rate limited; honor Retry-After
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unavailable:
      description: Temporary failure; retry original identity
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"